+++ changelog.d/7288.sdk
@@ -1,0 +1,10 @@
+Add `formattedText` or similar optional parameters in several methods:
+
+* RelationService:
+	* editTextMessage
+	* editReply
+	* replyToMessage
+* SendService:
+	* sendQuotedTextMessage
+
+This allows us to send any HTML formatted text message without needing to rely on automatic Markdown > HTML translation. All these new parameters have a `null` value by default, so previous calls to these API methods remain compatible.
