+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/model/relation/RelationService.kt
@@ -92,7 +92,7 @@ interface RelationService {
      * @param targetEvent The event to edit
      * @param msgType the message type
      * @param newBodyText The edited body in plain text
-     * @param newBodyTextFormatted The edited body with format
+     * @param newFormattedBodyText The edited body with format
      * @param newBodyAutoMarkdown true to parse markdown on the new body
      * @param compatibilityBodyText The text that will appear on clients that don't support yet edition
      */
@@ -100,7 +100,7 @@ interface RelationService {
             targetEvent: TimelineEvent,
             msgType: String,
             newBodyText: CharSequence,
-            newBodyTextFormatted: CharSequence? = null,
+            newFormattedBodyText: CharSequence? = null,
             newBodyAutoMarkdown: Boolean,
             compatibilityBodyText: String = "* $newBodyText"
     ): Cancelable
@@ -111,14 +111,14 @@ interface RelationService {
      * @param replyToEdit The event to edit
      * @param originalTimelineEvent the message that this reply (being edited) is relating to
      * @param newBodyText The plain text edited body (stripped from in reply to content)
-     * @param newBodyTextFormatted The formatted edited body (stripped from in reply to content)
+     * @param newFormattedBodyText The formatted edited body (stripped from in reply to content)
      * @param compatibilityBodyText The text that will appear on clients that don't support yet edition
      */
     fun editReply(
             replyToEdit: TimelineEvent,
             originalTimelineEvent: TimelineEvent,
             newBodyText: String,
-            newBodyTextFormatted: String? = null,
+            newFormattedBodyText: String? = null,
             compatibilityBodyText: String = "* $newBodyText"
     ): Cancelable
 
@@ -137,7 +137,7 @@ interface RelationService {
      * by the sdk into pills.
      * @param eventReplied the event referenced by the reply
      * @param replyText the reply text
-     * @param replyTextFormatted the reply text, formatted
+     * @param replyFormattedText the reply text, formatted
      * @param autoMarkdown If true, the SDK will generate a formatted HTML message from the body text if markdown syntax is present
      * @param showInThread If true, relation will be added to the reply in order to be visible from within threads
      * @param rootThreadEventId If show in thread is true then we need the rootThreadEventId to generate the relation
@@ -145,7 +145,7 @@ interface RelationService {
     fun replyToMessage(
             eventReplied: TimelineEvent,
             replyText: CharSequence,
-            replyTextFormatted: CharSequence? = null,
+            replyFormattedText: CharSequence? = null,
             autoMarkdown: Boolean = false,
             showInThread: Boolean = false,
             rootThreadEventId: String? = null
