+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/session/room/send/SendService.kt
@@ -69,7 +69,7 @@ interface SendService {
     fun sendQuotedTextMessage(
             quotedEvent: TimelineEvent,
             text: String,
-            formattedText: String?,
+            formattedText: String? = null,
             autoMarkdown: Boolean,
             rootThreadEventId: String? = null
     ): Cancelable
