+++ vector/src/main/java/im/vector/app/features/home/room/detail/composer/MessageComposerView.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2019 New Vector Ltd
+ * Copyright (c) 2022 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,177 +16,34 @@
 
 package im.vector.app.features.home.room.detail.composer
 
-import android.content.Context
-import android.net.Uri
 import android.text.Editable
-import android.util.AttributeSet
-import android.view.ViewGroup
 import android.widget.EditText
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.constraintlayout.widget.ConstraintSet
-import androidx.core.text.toSpannable
-import androidx.core.view.isInvisible
-import androidx.core.view.isVisible
-import androidx.transition.ChangeBounds
-import androidx.transition.Fade
-import androidx.transition.Transition
-import androidx.transition.TransitionManager
-import androidx.transition.TransitionSet
-import im.vector.app.R
-import im.vector.app.core.extensions.setTextIfDifferent
-import im.vector.app.databinding.ComposerLayoutBinding
 
-/**
- * Encapsulate the timeline composer UX.
- */
-class MessageComposerView @JvmOverloads constructor(
-        context: Context,
-        attrs: AttributeSet? = null,
-        defStyleAttr: Int = 0
-) : ConstraintLayout(context, attrs, defStyleAttr), MessageComposer {
+interface MessageComposerView {
 
-    interface Callback : ComposerEditText.Callback {
-        fun onCloseRelatedMessage()
-        fun onSendMessage(text: CharSequence)
-        fun onAddAttachment()
-        fun onExpandOrCompactChange()
-    }
+    val text: Editable?
+    val formattedText: String?
+    val editText: EditText
+    val emojiButton: ImageButton?
+    val sendButton: ImageButton
+    val attachmentButton: ImageButton
+    val composerRelatedMessageTitle: TextView
+    val composerRelatedMessageContent: TextView
+    val composerRelatedMessageImage: ImageView
+    val composerRelatedMessageActionIcon: ImageView
+    val composerRelatedMessageAvatar: ImageView
 
-    val views: ComposerLayoutBinding
+    var callback: PlainTextComposerLayout.Callback?
 
-    override var callback: Callback? = null
+    var isVisible: Boolean
 
-    private var currentConstraintSetId: Int = -1
+    fun collapse(animate: Boolean = true, transitionComplete: (() -> Unit)? = null)
+    fun expand(animate: Boolean = true, transitionComplete: (() -> Unit)? = null)
+    fun setTextIfDifferent(text: CharSequence?): Boolean
+    fun replaceFormattedContent(text: CharSequence)
 
-    private val animationDuration = 100L
-
-    override val text: Editable?
-        get() = views.composerEditText.text
-
-    override val formattedText: String? = null
-
-    override val editText: EditText
-        get() = views.composerEditText
-
-    override val emojiButton: ImageButton?
-        get() = views.composerEmojiButton
-
-    override val sendButton: ImageButton
-        get() = views.sendButton
-
-    override fun setInvisible(isInvisible: Boolean) {
-        this.isInvisible = isInvisible
-    }
-    override val attachmentButton: ImageButton
-        get() = views.attachmentButton
-    override val composerRelatedMessageActionIcon: ImageView
-        get() = views.composerRelatedMessageActionIcon
-    override val composerRelatedMessageAvatar: ImageView
-        get() = views.composerRelatedMessageAvatar
-    override val composerRelatedMessageContent: TextView
-        get() = views.composerRelatedMessageContent
-    override val composerRelatedMessageImage: ImageView
-        get() = views.composerRelatedMessageImage
-    override val composerRelatedMessageTitle: TextView
-        get() = views.composerRelatedMessageTitle
-    override var isVisible: Boolean
-        get() = views.root.isVisible
-        set(value) { views.root.isVisible = value }
-
-    init {
-        inflate(context, R.layout.composer_layout, this)
-        views = ComposerLayoutBinding.bind(this)
-
-        collapse(false)
-
-        views.composerEditText.callback = object : ComposerEditText.Callback {
-            override fun onRichContentSelected(contentUri: Uri): Boolean {
-                return callback?.onRichContentSelected(contentUri) ?: false
-            }
-
-            override fun onTextChanged(text: CharSequence) {
-                callback?.onTextChanged(text)
-            }
-        }
-        views.composerRelatedMessageCloseButton.setOnClickListener {
-            collapse()
-            callback?.onCloseRelatedMessage()
-        }
-
-        views.sendButton.setOnClickListener {
-            val textMessage = text?.toSpannable() ?: ""
-            callback?.onSendMessage(textMessage)
-        }
-
-        views.attachmentButton.setOnClickListener {
-            callback?.onAddAttachment()
-        }
-    }
-
-    override fun replaceFormattedContent(text: CharSequence) {
-        setTextIfDifferent(text)
-    }
-
-    override fun collapse(animate: Boolean, transitionComplete: (() -> Unit)?) {
-        if (currentConstraintSetId == R.layout.composer_layout_constraint_set_compact) {
-            // ignore we good
-            return
-        }
-        currentConstraintSetId = R.layout.composer_layout_constraint_set_compact
-        applyNewConstraintSet(animate, transitionComplete)
-        callback?.onExpandOrCompactChange()
-    }
-
-    override fun expand(animate: Boolean, transitionComplete: (() -> Unit)?) {
-        if (currentConstraintSetId == R.layout.composer_layout_constraint_set_expanded) {
-            // ignore we good
-            return
-        }
-        currentConstraintSetId = R.layout.composer_layout_constraint_set_expanded
-        applyNewConstraintSet(animate, transitionComplete)
-        callback?.onExpandOrCompactChange()
-    }
-
-    override fun setTextIfDifferent(text: CharSequence?): Boolean {
-        return views.composerEditText.setTextIfDifferent(text)
-    }
-
-    private fun applyNewConstraintSet(animate: Boolean, transitionComplete: (() -> Unit)?) {
-        // val wasSendButtonInvisible = views.sendButton.isInvisible
-        if (animate) {
-            configureAndBeginTransition(transitionComplete)
-        }
-        ConstraintSet().also {
-            it.clone(context, currentConstraintSetId)
-            it.applyTo(this)
-        }
-        // Might be updated by view state just after, but avoid blinks
-        // views.sendButton.isInvisible = wasSendButtonInvisible
-    }
-
-    private fun configureAndBeginTransition(transitionComplete: (() -> Unit)? = null) {
-        val transition = TransitionSet().apply {
-            ordering = TransitionSet.ORDERING_SEQUENTIAL
-            addTransition(ChangeBounds())
-            addTransition(Fade(Fade.IN))
-            duration = animationDuration
-            addListener(object : Transition.TransitionListener {
-                override fun onTransitionEnd(transition: Transition) {
-                    transitionComplete?.invoke()
-                }
-
-                override fun onTransitionResume(transition: Transition) {}
-
-                override fun onTransitionPause(transition: Transition) {}
-
-                override fun onTransitionCancel(transition: Transition) {}
-
-                override fun onTransitionStart(transition: Transition) {}
-            })
-        }
-        TransitionManager.beginDelayedTransition((parent as? ViewGroup ?: this), transition)
-    }
+    fun setInvisible(isInvisible: Boolean)
 }
